requirements/development.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/development.txt requirements/development.in
#
amqp==2.3.2               # via kombu
appnope==0.1.0            # via ipython
asgiref==2.3.2            # via channels
asn1crypto==0.24.0        # via cryptography
async-timeout==3.0.0      # via asgiref
attrs==18.1.0             # via automat, twisted
autobahn==18.8.1          # via daphne
automat==0.7.0            # via twisted
autopep8==1.3.5
babel==2.6.0              # via flower
backcall==0.1.0           # via ipython
bcrypt==3.1.4             # via paramiko
billiard==3.5.0.4         # via celery
boto3==1.7.70
botocore==1.10.70         # via boto3, s3transfer
celery==4.2.1
certifi==2018.4.16        # via requests
cffi==1.11.5              # via bcrypt, cryptography, pynacl
channels==2.1.3
chardet==3.0.4            # via requests
constantly==15.1.0        # via twisted
coverage==4.5.1
cryptography==2.3         # via paramiko
daphne==2.2.2             # via channels
decorator==4.3.0          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid, social-auth-core
django-braces==1.13.0     # via django-rest-framework-social-oauth2
django-cacheops==4.0.7
django-celery-beat==1.1.1
django-constance[database]==2.2.0
django-cors-headers==2.4.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-extensions==2.1.0
django-filter==2.0.0
django-hstore==1.4.2      # via djangorestframework-hstore
django-jenkins==0.110.0
django-model-utils==3.1.2
django-money==0.14.2
django-oauth-toolkit==1.2.0  # via django-rest-framework-social-oauth2
django-object-actions==1.0.0
django-pdb==0.6.2
django-picklefield==1.0.0  # via django-constance
django-redis==4.9.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.6.6
django-taggit==0.23.0
django-test-plus==1.1.1
django-test-without-migrations==0.6
django==2.1
djangorestframework-gis==0.13
djangorestframework-hstore==1.3
djangorestframework==3.8.2
docopt==0.6.2             # via ptpython
docutils==0.14            # via botocore
elpy==1.23.0
fabric3==1.14.post1
factory-boy==2.11.1
faker==0.8.17             # via factory-boy
flake8==3.5.0
flower==0.9.2
freezegun==0.3.10
funcy==1.10.3
hyperlink==18.0.0         # via twisted
idna==2.7                 # via cryptography, hyperlink, requests
importmagic==0.1.7
incremental==17.5.0       # via twisted
inflection==0.3.1
ipdb==0.11
ipython-genutils==0.2.0
ipython==6.5.0
isort==4.3.4
jedi==0.12.1
jmespath==0.9.3           # via boto3, botocore
kombu==4.2.1              # via celery
mccabe==0.6.1             # via flake8
oauthlib==2.1.0           # via django-oauth-toolkit, requests-oauthlib, social-auth-core
paramiko==2.4.1           # via fabric3
parso==0.3.1              # via jedi
pep8==1.7.1
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.2.0
prompt-toolkit==1.0.15    # via ipython, ptpython
psycopg2==2.7.5
ptipython==1.0.0
ptpython==0.41
ptyprocess==0.6.0         # via pexpect
py-moneyed==0.7.0         # via django-money
pyasn1==0.4.4             # via paramiko
pycodestyle==2.3.1        # via autopep8, flake8
pycparser==2.18           # via cffi
pyflakes==1.6.0           # via flake8
pygments==2.2.0           # via ipython, ptpython
pyhamcrest==1.9.0         # via twisted
pyjwt==1.6.4              # via social-auth-core
pynacl==1.2.1             # via paramiko
python-dateutil==2.7.3    # via botocore, faker, freezegun
python3-openid==3.1.0     # via social-auth-core
pytz==2018.5
redis==2.10.6
requests-oauthlib==1.0.0  # via social-auth-core
requests==2.19.1          # via django-oauth-toolkit, requests-oauthlib, social-auth-core, vk
rest-social-auth==1.3.1
rope==0.10.7
s3transfer==0.1.13        # via boto3
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via autobahn, automat, bcrypt, cryptography, django-cacheops, django-coverage-plugin, django-extensions, djangorestframework-gis, fabric3, faker, freezegun, prompt-toolkit, pyhamcrest, pynacl, python-dateutil, social-auth-app-django, social-auth-core, traitlets, txaio
social-auth-app-django==2.1.0
social-auth-core==1.7.0   # via rest-social-auth, social-auth-app-django
sqlparse==0.2.4           # via django-debug-toolbar
tblib==1.3.2
text-unidecode==1.2       # via faker
tornado==5.1              # via flower
traitlets==4.3.2          # via ipython
twisted==18.7.0           # via daphne
txaio==18.7.1             # via autobahn
urllib3==1.23             # via requests
vine==1.1.4               # via amqp
vk==2.0.2
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.14.1
yapf==0.22.0
zope.interface==4.5.0     # via twisted
